+++ app/Services/ProductService.php
@@ -9,6 +9,7 @@ use App\Models\Category;
 use App\Models\Product;
 use App\Models\Shop;
 use App\Repositories\Contracts\ProductInterface;
+use Illuminate\Support\Facades\Auth;
 // use Illuminate\Support\Facades\Storage;
 
 class ProductService
@@ -50,8 +51,8 @@ class ProductService
     {
         $category = Category::findOrFail($request->category_id);
         $data = $this->translationService->CategoryLangs($category->id);
-        if ($data['unavailable_langs_count'] > 0 or ! empty($data['unavailable_langs'])) {
-            throw new \Exception("You cannot use this category for now because it is Missing translation, please waite for all translations to be available", 500);
+        if ($data['available_langs_count'] == 0 or empty($data['available_langs'])) {
+            throw new \Exception("You cannot use this category for now because it is Missing translation, please waite for any translation to be available.", 500);
         }
         return $this->productRepository->store($request);
     }
@@ -59,7 +60,7 @@ class ProductService
     public function update(UpdateProductRequest $request, int $id)
     {
         Category::findOrFail($request->category_id);
-        $product = Product::findOrFail($id);
+        $product = Product::whereKey($id)->where('user_id', Auth::id())->firstOrFail();
         return $this->productRepository->update($request, $product);
     }
 
@@ -71,14 +72,35 @@ class ProductService
 
     public function unban(int $id)
     {
-        $product = Product::findOrFail($id);
+        $product = Product::withTrashed()->findOrFail($id);
         return $this->productRepository->unban($product);
+    }
+
+    public function deleteMine(int $id)
+    {
+        $product = Product::whereKey($id)->where('user_id', Auth::id())->firstOrFail();
+
+        // هذا المنطق مُنَفى للإستخدام التجارى   $product->orders()->exists()
+        if ($product->orders()->exists() or $product->carts()->exists()) {
+            throw new \Exception(
+                'Can not delete the product because it has related carts or orders. They have get deleted first.'
+            );
+        }
+        // Storage::deleteDirectory('products/'.$product->id.'/');
+
+        return $this->productRepository->delete($product);
     }
 
     public function delete(int $id)
     {
         $product = Product::findOrFail($id);
-
+        
+        // هذا المنطق مُنَفى للإستخدام التجارى  $product->orders()->exists()
+        if ($product->orders()->exists() or $product->carts()->exists()) {
+            throw new \Exception(
+                'Can not delete the product because it has related carts or orders. They have get deleted first.'
+            );
+        }
         // Storage::deleteDirectory('products/'.$product->id.'/');
 
         return $this->productRepository->delete($product);
